Created separate repository for data fetched from network. Moved all network query files to new packages.
